https://github.com/itext/itext7/releases/tag/7.1.14

* release/7.1.14:
  [RELEASE] 7.1.14
  Add unit tests for SvgCoordinateUtils
  Clean up manual porting for java's Matcher
  Restore lost tests introduced in DEVSIX-4682
  Support 'patternTransform' attribute for SVG pattern element
  Add support to utf-8 image format in svg
  Support 'viewbox' and `preserveAspectRatio' attribute for SVG pattern element
  Add test on second time signing document
  Add tests for SVG moveTo followed by multiple pairs
  Refactor text renderers sequence word wrapping logic
  Separate text sequence reset logic for robustness
  Support overflow-wrap: break-word/anywhere in case of the not fitting single glyph
  Remove floats skipping logic in span-wrapping: now floats break text sequence
  Support word-break property
  Support minMaxWidth counting for span- and special scripts wrapping
  Change condition to stop TextRenderer#layout when a width limit is reached
  Text elements wrapping support
  Add missing copyright headers
  Fix dummy log messages when we are processing xlink href with data
  Improve test coverage for StopSvgNodeRenderer class
  Add missing copyright headers
  Add PdfObjectReleaseTest class
  Add tests with all options of preserveAspectRation in svg element
  Add missing copyright headers
  Add tests with relative units in attributes of svg elements
  Add tests for absolutely positioned elements added to Canvas
  Fix javadoc
  Add stroke-width into inheritable set of svg attributes, update cmp files
  Upgrade BouncyCastle version to 1.67
  Fix javadoc warnings
  Add KeepTogetherTest#keepTogetherInDivWithKidsFloatTest
  Support 'objectBoundingBox' value for SVG pattern element 'patternContentUnits' and 'patternUnits'
  Unify the resource folder for HighlightItemsTest class
  Add form field borders support
  Add JavaDocs to unsupported AreaBreakRenderer methods
  Add support of pages target-counter(s). Finalize whole target-counter logic
  Add missing copyright headers
  Implement non-page(s) target-counter(s)
  Matcher tests
  Add initial support of SVG pattern element
  Configure dependency-check-maven to connect to a central database
  Support overflow-x for special scripts
  Support css overflow-wrap property
  Update veraPDF version in pdftest module
  Update printed links in CompareTool
  Update CONTRIBUTING.md with latest information and links
  Update BouncyCastle version to 1.66
  Divide the CssUtils class into several classes
  Move HighlightItemsTest from samples-internal repository
  Update slf4j version from 1.7.13 to 1.7.30
  Delete TODO related to clossed ticket
  Add intergration tests reproducing keep-together causing content overlap
  Make mutable final sets unmodifable
  Add XXEVulnerabilityTest class
  Add new tests for PdfTextExtractor
  Speed up PdfPagesTest#randomNumberPagesTest()
  Add missing copyright headers
  Implement target-counter for page
  Revert "Update junit version"
  Update junit version
  Update Java version from 1.7 to 1.8
  Remove TODO related to closed DEVSIX-2785
  Get rid of TODOs not related to Jira tickets in SignatureUtil
  BlockRenderer keep together for clear child element fix
  Add tests for keep-together elements in collaboration with floats wrapped into non-floating element
  Add intedration tests for svg pattern element
  Add missing copyright headers
  Add support of font-relative units for the font-size property
  Fix NPE in CanvasTag#getActualText() in case properties are missing
  Improve javadocs in svg and xml-parser modules
  Update japicmp version to 0.14.4
  Add missing copyright headers
  Improving test coverage for PdfFont class
  Fix javadoc warnings
  Remove package-info.java
  Fix JavaDoc warnings in several classes
  Fix javadoc warnings in PdfExplicitDestination and in PdfExplicitRemoteGoToDestination
  Remove itext5 verison tag from root pom file
  [AFTER RELEASE] 7.1.13
  Allow fetching of OTF language records by tag
  Add tests on default font-size and inheritance inside TSpan element
  Add missing copyright headers
  Fix overflow in OTF readCoverageFormat() DEVSIX-4642

+++ kernel/src/main/java/com/itextpdf/kernel/numbering/AlphabetNumbering.java
@@ -69,8 +69,8 @@ public class AlphabetNumbering {
 
         number--;
         int bytes = 1;
-        int start = 0;
-        int symbols = cardinality;
+        long start = 0;
+        long symbols = cardinality;
 
         while (number >= symbols + start) {
             bytes++;
@@ -78,10 +78,10 @@ public class AlphabetNumbering {
             symbols *= cardinality;
         }
 
-        int c = number - start;
+        long c = number - start;
         char[] value = new char[bytes];
         while (bytes > 0) {
-            value[--bytes] = alphabet[c % cardinality];
+            value[--bytes] = alphabet[(int) (c % cardinality)];
             c /= cardinality;
         }
